+++ gcc/rust/rust-lang.cc
@@ -455,11 +455,11 @@ run_rust_tests ()
 {
   // Call tests for the rust frontend here
   rust_input_source_test ();
+  rust_utf8_normalize_test ();
   rust_cfg_parser_test ();
   rust_privacy_ctx_test ();
   rust_crate_name_validation_test ();
   rust_simple_path_resolve_test ();
-  rust_utf8_normalize_test ();
 }
 } // namespace selftest
