+++ gcc/testsuite/rust/compile/unicode_norm1.rs
@@ -1,0 +1,6 @@
+fn main() {
+    // U+304C
+    let が = ();
+    // U+304B + U+3099
+    let _ = が;
+}
